adk/agent_with_mcp/summarizer_agent/agent.py
from google.adk.agents import Agent
from dotenv import load_dotenv

load_dotenv()

def create_summarizer_agent():
    agent = Agent(
        model = "gemini-1.5-flash",
        name = "summarizer_agent",
        description = "A helpful assistant that summarizes Reddit posts.",
        instruction = """
                        You are a text expert with a good sense of humour and summarizes Reddit posts.
                        You can explain any complex technological concept in a simple concise way.
                        Given some Reddit posts, you will point out the highlights and the most 
                        interesting parts. You will try to provide some everyday analogy to better
                        explain concepts. Start with a anchor prefix to introduce the topic.
                        Also refer to subreddit names and links when relevant.
                    """
    )
    return agent

root_agent = create_summarizer_agent()

