pyproject.toml
[project]
name = "python-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "google-adk>=0.3.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "praw>=7.8.1",
]
